improvements

+++ Classes/AppDelegate.cpp
@@ -46,10 +46,11 @@ bool AppDelegate::applicationDidFinishLaunching() {
     }
 
     // set FPS. the default value is 1.0/60 if you don't call this
-    director->setAnimationInterval((float)1.0 / 60);
+    director->setAnimationInterval(1.0f / 60);
 
     // Set the design resolution
     glView->setDesignResolutionSize(designResolutionSize.width, designResolutionSize.height, ResolutionPolicy::NO_BORDER);
+
     auto frameSize = glView->getFrameSize();
     // if the frame's height is larger than the height of large size.
     if (frameSize.height > mediumResolutionSize.height) {
